# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: FMS Log Nexus API
  description: |
    Centralized logging and monitoring platform API for FMS Log Nexus.
    
    ## Authentication
    
    The API supports two authentication methods:
    - **JWT Bearer Token**: For dashboard users and interactive sessions
    - **API Key**: For service accounts and job integrations
    
    ## Rate Limiting
    
    - Log ingestion: 10,000 requests/minute per API key
    - Dashboard queries: 1,000 requests/minute per user
    
  version: 1.0.0
  contact:
    name: FMS Log Nexus Support
    email: support@company.com
  license:
    name: Proprietary
servers:
  - url: https://lognexus.company.com/api
    description: Production server
  - url: https://localhost:5001/api
    description: Development server
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate user
      description: Authenticate with username and password to obtain JWT tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or expired refresh token
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token for user authentication

    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service authentication
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
            traceId:
              type: string

    # ========== Auth ==========
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          default: Bearer
        expiresIn:
          type: integer
          description: Token expiration in seconds
        user:
          $ref: '#/components/schemas/User'
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
        expiresIn:
          type: integer
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
        role:
          type: string

    # ========== Logs ==========
...